refactor(upgrade): move shared output helpers to utils and add happy-path test
count_upgrade_categories and format_upgrade_summary are used by both
upgrade.rs and activate.rs, so they belong in utils/upgrade_output.rs
rather than owned by the upgrade command. Adds a dry-run integration test
that exercises the full output path with an actual version change.

Forge-Agent: implementation-worker
