+++ cli/flox/src/utils/mod.rs
@@ -18,6 +18,7 @@ pub mod openers;
 pub mod search;
 pub mod tracing;
 pub mod update_notifications;
+pub mod upgrade_output;
 
 pub static TERMINAL_STDERR: LazyLock<Mutex<Stderr>> =
     LazyLock::new(|| Mutex::new(std::io::stderr()));
